Tests for grouping candu graphs by tags
